Support the official BF16 Qwen3.8-27B checkpoint
The official Qwen/Qwen3.8-27B release is dense BF16 and bundles a vision
tower in the same 18 shards as the text model. Mapping it needs no new
architecture, but it does need explicit coverage accounting so a partial
load cannot pass silently, and it needs a host-only audit path so a
52 GiB checkpoint can be verified without an accelerator toolkit.

Split the Qwen loader along the device boundary: mapping, shard
descriptors, coverage and host materialization move to
core/qwen_weight_map.cpp and link only pocket_core, while device
residency and uploads stay in engine/qwen_weights.cpp. Add QwenCoverage,
which classifies every index entry as mapped, deliberately ignored
vision weights, or unexpected, and require_full_coverage() to reject
anything unaccounted for.

Audit modes no longer bind a device, so --qwen-audit --tp-world 4 works
on a host with no CUDA runtime and fewer than four visible GPUs.

Verified on the real checkpoint: 18 shard headers parse, 866 text and
333 vision tensors on every TP4 rank with zero unexpected entries,
12.796 GiB resident per rank, and identical rank metadata. CUDA
generation is unvalidated and documented as such.
